requirements.txt
altair==5.4.1
attrs==24.2.0
Automat==22.10.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.6.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.4.0
Protego==0.3.1
protobuf==5.28.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.7.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rich==13.8.0
rpds-py==0.20.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==72.2.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.34
streamlit==1.38.0
tenacity==8.5.0
tldextract==5.1.2
toml==0.10.2
tornado==6.4.1
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
watchdog==4.0.2
zope.interface==7.0.1
